refactor(cli): move getPermission function to transformer.ts
